requirements.txt
Django>=2.1.0,<2.2.0
djangorestframework>3.9.1<3.4.0
flake8>=3.6.0,<3.7.0
psycopg2<=2.8.4 
pillow